*: upgrade to 3.2
Fix errors when built with https://github.com/meh/rust-ffmpeg-sys/pull/35.

+++ src/codec/encoder/encoder.rs
@@ -1,6 +1,6 @@
 use std::ops::{Deref, DerefMut};
 
-use libc::c_int;
+use libc::{c_int, int64_t};
 use codec::Context;
 use ::{Error, Rational, media};
 use super::{video, audio, subtitle};
@@ -70,13 +70,13 @@ impl Encoder {
 
 	pub fn set_bit_rate(&mut self, value: usize) {
 		unsafe {
-			(*self.as_mut_ptr()).bit_rate = value as c_int;
+			(*self.as_mut_ptr()).bit_rate = value as int64_t;
 		}
 	}
 
 	pub fn set_max_bit_rate(&mut self, value: usize) {
 		unsafe {
-			(*self.as_mut_ptr()).rc_max_rate = value as c_int;
+			(*self.as_mut_ptr()).rc_max_rate = value as int64_t;
 		}
 	}
